mm: small bugfix in memutils
replace qword with dword because right now we
operate on 32 bit kernel

+++ include/mm/mem_utils.h
@@ -3,7 +3,7 @@
 
 #include <main/types.h>
 
-static inline void* memcpy(void* dst, const void* src, u64 n) {
+static inline void* memcpy(void* dst, const void* src, u32 n) {
     asm volatile(
         "cld\n\t"
         "rep movsb"
@@ -14,7 +14,7 @@ static inline void* memcpy(void* dst, const void* src, u64 n) {
     return dst;
 }
 
-static inline void* memset(void* dst, int c, u64 n) {
+static inline void* memset(void* dst, int c, u32 n) {
     asm volatile(
         "cld\n\t"
         "rep stosb"
@@ -25,7 +25,7 @@ static inline void* memset(void* dst, int c, u64 n) {
     return dst;
 }
 
-static inline int memcmp(const void* a, const void* b, u64 n) {
+static inline int memcmp(const void* a, const void* b, u32 n) {
     const u8* p1 = (const u8*)a;
     const u8* p2 = (const u8*)b;
     for (u64 i = 0; i < n; i++) {
@@ -35,7 +35,7 @@ static inline int memcmp(const void* a, const void* b, u64 n) {
     return 0;
 }
 
-static inline void* memmove(void* dst, const void* src, u64 n) {
+static inline void* memmove(void* dst, const void* src, u32 n) {
     u8* d = (u8*)dst;
     const u8* s = (const u8*)src;
     if (d < s || d >= s + n) {
